Moved CustomIdentityContentCodec to test scope
