+++ servicetalk-encoding-api/build.gradle
@@ -30,6 +30,4 @@ dependencies {
     testImplementation "org.hamcrest:hamcrest-library:$hamcrestVersion"
 
     testRuntimeOnly "org.junit.jupiter:junit-jupiter-engine:$junit5Version"
-
-    testImplementation testFixtures(project(":servicetalk-encoding-api-internal"))
 }
