+++ servicetalk-encoding-api/src/test/java/io/servicetalk/encoding/api/IdentityContentCodecTest.java
@@ -16,9 +16,6 @@
 
 package io.servicetalk.encoding.api;
 
-import io.servicetalk.encoding.api.internal.CustomIdentityContentCodec;
-import io.servicetalk.encoding.api.internal.NoopContentCodec;
-
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
